Update test setup to pull a student associated to a school that is also linked to an LEA
The setup could accidentally pull in one not linked so then the subsequent api call
in the test would return no results and fail. Now with it linked that will not happen.
Adding this null check is similar to other test setups as well. Adding this allowed the this
test and another test to no longer be ignored.