[ODS-4835] Code Generation for Resource Links to some Base Aggregates Fail (#436)
* Removing unused references to base property mappings to fix issues with link generation for GeneralStudentProgramAssociation links.

* Updated approval tests.

* Remove now unused methods

* Skip currently failing composite test that cannot be solved

* Update test setup to pull a student associated to a school that is also linked to an LEA

The setup could accidentally pull in one not linked so then the subsequent api call
in the test would return no results and fail. Now with it linked that will not happen.
Adding this null check is similar to other test setups as well. Adding this allowed the this
test and another test to no longer be ignored.

* Update hql alias to match latest changes

Test was previously ignored and now brought back in and the alias in use has changed

* Update query to join to edorg mapping table and pull from a specific school to find a student

This should make it so the query cannot pull from a school we do not want it to

* Fix alias to match the one declared above

Co-authored-by: Benjamin Meyers <blmeyers@users.noreply.github.com>
Co-authored-by: Ben Tillman <ben@headspring.com>