Added a JsonIgnore property to the ResourceProperty along with the logic for identifying when a property should be ignored from the JSON (and related API metadata).
Incorporated use of the JsonIgnore property in the Resources.mustache template to generate a JsonIgnore attribute in lieu of a Displayname attribute on resource properties used for JSON serialization.

Incorporated use of the JsonIgnore property into the resource extension method NonReferencedProperties which is used by the OpenApiMetadataDefinitionsFactory class to produce property-level Open API metadata.